Made file upload more robust
Uploaded files are now named a random string of letters and digits, and
any (rare) race conditions should have been avoided as the script gets
a file hook on the file whilst testing for its existance. I say race
conditions are rare as it would require two random 32 character strings to
be the same!