Added Ansible playbooks to automate database backup and restoration
Added Ansible playbooks which copy all images as well as MongoDB
backups across from the remote to the local as well as the other
way around. At the moment, the database is only backed up when I
run the script. Ultimately, the database backup should be automated
with cron, with the backup script just copying the files across to
the local destination.

I also moved a few files around to make the repository cleaner.