Added "Advanced Measurements" and code checking to the upload page
"Advanced Measurements" are provided as a hidden panel, designed
for users with superior quantification equipment. The hope is that
these users  will be able to submit photos with these measurements
to improve the prediction of the measurements for users who do not
have access to more advanced measurement apparatus.

Also implemented is a check for a correct upload code, and a prompt
to Tweet about participation if an upload was successful.