Bug fix to prevent database update when registration errors
Now the send e-mail function is wrapped in a try-except structure
to catch errors.

I also fixed the provisioning of the Vagrant box using Ansible,
as there was a problem with the chcon section. I think this is
due to there being different SELinux policies on the Vagrant box
and on the production server.