Got the server code working with the image analysis code
OpenCV wasn't working when run in a Python sub-interpreter, as
mod_wsgi does. Correcting the Apache configuration file sorted
this. I also modified routes.py and the ExperimentPage.html
files to give the user feedback on the image they submitted.
Finally, there was a bug in the image analysis code which meant
that handles were sometimes not properly being retreived from
images. The code I've put in place isn't ideal, but it fixes the
problem at least on my test image. A proper test suite is required,
but for that I ideally need a collection of images with which the
current code fails. I can only get that when users start submitting
images to me.