Added a details list of images which failed analysis
The list is in the global variables collection. If an image fails
image analysis by not having a detectable calibration window, then
an entry is added to this list with some of the details of the
image. The idea behind this is to track what causes the image
analysis code to fail, which will hopefully lead to code which is
more robust.

The original image filename is now also saved regardless of whether
the image analysis was successful or not. This is a way of getting
around the 'feature' of Apple phones stripping the EXIF data
from images, as I think the date and time the image was created is
usually saved in the file name.