Changed layout of Experiment page and fixed 2 image analysis bugs
The experiment page was changed around to remove distracting information and emphasize the Upload form when accessing the site on mobile.

Some images (specifically, images taken from an Apple device and uploaded from a computer) were causing an error when they were added to the database as they could not be UTF-8 encoded. I found this was due to the MakerNote entry in the EXIF dictionary containing invalid characters. There is now a test for this, and if the string cannot be encoded, is removed from the dictionary. In the future it might make sense to just remove the offending characters and keep the rest. I do not anticipate this being a problem when uploading images directly from Apple products and they strip the EXIF data before upload.

The need for a dynamic Gaussian blur kernel has been highlighted. The Gaussian blur kernel size was initially set as a magic number of 21. However, if a small image is uploaded this setting is too high. I've put in a temporary fix which uses a kernel size of 5 if the image width is below 200 pixels, but ideally the kernel size will be dynamic.